pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name="caen-tools"
version="2.2.2"
requires-python=">=3.11"
dependencies=[
    "pyzmq==25.1.1",
    "caen-setup @ git+https://github.com/caenHV/Setup.git@v1.2.2",
    "psycopg2-binary==2.9.9",
]

[project.scripts]
caen_device_backend = "caen_tools.DeviceBackend.main:main"
caen_monitor = "caen_tools.MonitorService.monitor:main"
caen_webserver = "caen_tools.WebService.ws:main"
caen_system_check = "caen_tools.SystemCheck.check:main"

[project.optional-dependencies]
webservice=[
    "fastapi==0.104.1",
    "typing_extensions==4.8.0",
    "uvicorn==0.23.2",
    "typing-inspect==0.9.0",
    "sse-starlette==2.1.3",
    "async-lru==2.0.4",
]
